update to add more model seperation
