+++ conf/model/default.yaml
@@ -5,4 +5,4 @@ defaults:
 # By default we turn off recursive instantiation, allowing the user to instantiate themselves at the appropriate times.
 _recursive_: false
 
-_target_: torch_points3d.models.base_model.PointCloudBaseModel
+_target_: torch_points3d.tasks.base_model.PointCloudBaseModule
