+++ conf/model/segmentation/sparseconv3d/ResUNet32.yaml
@@ -2,39 +2,40 @@
 defaults:
   - /model/segmentation/default
 
-backbone:
-  _target_: torch_points3d.applications.sparseconv3d.SparseConv3d
-  backend: torchsparse
+model: 
+  backbone:
+    _target_: torch_points3d.applications.sparseconv3d.SparseConv3d
+    backend: torchsparse
 
-  config:
-    define_constants:
-      in_feat: 32
-      block: ResBlock # Can be any of the blocks in modules/MinkowskiEngine/api_modules.py
-    down_conv:
-      module_name: ResNetDown
-      block: block
-      N: [ 0, 1, 2, 2, 3 ]
-      down_conv_nn:
-        [
-          [ FEAT, in_feat ],
-          [ in_feat, in_feat ],
-          [ in_feat, 2*in_feat ],
-          [ 2*in_feat, 4*in_feat ],
-          [ 4*in_feat, 8*in_feat ],
-        ]
-      kernel_size: 3
-      stride: [ 1, 2, 2, 2, 2 ]
-    up_conv:
-      block: block
-      module_name: ResNetUp
-      N: [ 1, 1, 1, 1, 0 ]
-      up_conv_nn:
-        [
-          [ 8*in_feat, 4*in_feat ],
-          [ 4*in_feat + 4*in_feat, 4*in_feat ],
-          [ 4*in_feat + 2*in_feat, 3*in_feat ],
-          [ 3*in_feat + in_feat, 3*in_feat ],
-          [ 3*in_feat + in_feat, 3*in_feat ],
-        ]
-      kernel_size: 3
-      stride: [ 2, 2, 2, 2, 1 ]
+    config:
+      define_constants:
+        in_feat: 32
+        block: ResBlock # Can be any of the blocks in modules/MinkowskiEngine/api_modules.py
+      down_conv:
+        module_name: ResNetDown
+        block: block
+        N: [ 0, 1, 2, 2, 3 ]
+        down_conv_nn:
+          [
+            [ FEAT, in_feat ],
+            [ in_feat, in_feat ],
+            [ in_feat, 2*in_feat ],
+            [ 2*in_feat, 4*in_feat ],
+            [ 4*in_feat, 8*in_feat ],
+          ]
+        kernel_size: 3
+        stride: [ 1, 2, 2, 2, 2 ]
+      up_conv:
+        block: block
+        module_name: ResNetUp
+        N: [ 1, 1, 1, 1, 0 ]
+        up_conv_nn:
+          [
+            [ 8*in_feat, 4*in_feat ],
+            [ 4*in_feat + 4*in_feat, 4*in_feat ],
+            [ 4*in_feat + 2*in_feat, 3*in_feat ],
+            [ 3*in_feat + in_feat, 3*in_feat ],
+            [ 3*in_feat + in_feat, 3*in_feat ],
+          ]
+        kernel_size: 3
+        stride: [ 2, 2, 2, 2, 1 ]
