+++ torch_points3d/models/segmentation/base_model.py
@@ -1,16 +1,35 @@
-from torch_points3d.models.base_model import PointCloudBaseModel
-import torch.nn.functional as F
+from typing import Any, Dict, Optional, Tuple, Union
+from omegaconf import DictConfig
+
 import torch.nn as nn
+import torch.nn.functional as F
 
-# Overrides PointCloudBaseModel to create segmentation-specific models
+from torch_points3d.core.instantiator import Instantiator
+from torch_points3d.models.base_model import PointCloudBaseModel
+
+
 class SegmentationBaseModel(PointCloudBaseModel):
+    def __init__(self, instantiator: Instantiator, num_classes: int, backbone: DictConfig):
+        super().__init__(instantiator)
+        
+        print(backbone)
+        self.backbone = self.instantiator.backbone(backbone)  
+             
+        self.head = nn.Sequential(nn.Linear(self.backbone.output_nc, num_classes))
 
-    # Most segmentation models will share the same head and criterion
-    # Note that we don't even have to create a sparseconv3d subclass of PointCloudBaseModel
-    # because the actual training logic is contained all within this and PointCloudBaseModel
-    # so all we need to do is define a `backend` (in this case sparseconv3d)
-    def _build_model(self, backbone_cfg):
-        super()._build_model(backbone_cfg)
+    def set_input(self, data):
+        self.batch_idx = data.batch.squeeze()
+        self.input = data
+        if data.y is not None:
+            self.labels = data.y
+        else:
+            self.labels = None
+    
+    def forward(self):
+        features = self.backbone(self.input).x
+        logits = self.head(features)
+        self.output = F.log_softmax(logits, dim=-1)
 
-        self.head = nn.Sequential(nn.Linear(self.backbone.output_nc, self.num_classes))
-        self.criterion = F.nll_loss
+        # only compute loss if loss is defined and the dset has labels
+        if self.labels is not None and self.criterion is not None:
+            return self.criterion(self.output, self.labels)
